% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{CleanJSTOR_df}
\alias{CleanJSTOR_df}
\title{CleanJSTOR_df}
\usage{
CleanJSTOR_df(df, Titles = FALSE)
}
\arguments{
\item{df}{A dataframe in the JSTOR_df format (see: JSTOR_df())}

\item{Titles}{whether to clean titles (instead of abstracts). Default = FALSE.}
}
\value{
A JSTOR_df dataframe containing informations about: Journal, Publisher, Title, Year, Abstract and Language
}
\description{
This cleans a JSTOR_df, excluding repeated name papers, non-english papers, NAs (in titles or abstracts)
}
\examples{
CleanJSTOR_df(df)


}
