% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{compareJSTOR_dfs}
\alias{compareJSTOR_dfs}
\title{compareJSTOR_dfs}
\usage{
compareJSTOR_dfs(df1, df2, y, x, legend, comparisonType)
}
\arguments{
\item{df1}{A dataframe in the JSTOR_df format (see: JSTOR_df())}

\item{df2}{A dataframe in the JSTOR_df format (see: JSTOR_df())}

\item{y}{Determines the size of the ranking being observed for each year}

\item{x}{Determines the number of times a bigram has to be part of the rank to
be included in the dataframe or plot}

\item{legend}{A vector of charcters to be used as legends matching
the number of dfs (2): c("legend1", "legend2")}

\item{comparisonType}{Which type of caomparison to make. Acceptable entries: "Words", "Bigrams", "Trigrams" and "Journal".}
}
\value{
A ggplot2 object
}
\description{
Creates dataframe or plot with information regarding top bigrams that appear at least
for a determined time period in the top.
}
\examples{
compareJSTOR_dfs(dfx, dfk, 4, 4, c("legendx", "legendk", comparisonType = "Journal")


}
