% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyJSTOR.R
\name{JSTOR_df}
\alias{JSTOR_df}
\title{JSTOR_df}
\usage{
JSTOR_df(filePath)
}
\arguments{
\item{filePath}{}
}
\value{
A dataframe containing informations about: Journal, Publisher, Title, Year, Abstract and Language
}
\description{
This function transforms all files in a folder with JSTOR's DFR metada into a tidy dataframes
}
\examples{
JSTOR_df("C:/Users/.../Documents/metadata/")


}
