% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{JSTORplotJournals}
\alias{JSTORplotJournals}
\title{JSTORplotJournals}
\usage{
JSTORplotJournals(df, y, x, output = "plot")
}
\arguments{
\item{y}{Determines the size of the ranking being observed for each year}

\item{x}{Determines the number of times a bigram has to be part of the rank to
be included in the dataframe or plot}

\item{output}{Whether to output a dataframe or a plot. Acceptable entries: "plot" and "dataframe".}

\item{dfEco}{A dataframe in the JSTOR_df format (see: JSTOR_df())}
}
\value{
A ggplot2 object or a dataframe
}
\description{
Creates dataframe or plot with information regarding top Journal that appear at least
for a determined time period in the top.
}
\examples{


}
