% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{JSTORplotVocabCount}
\alias{JSTORplotVocabCount}
\title{JSTORplotVocabCount}
\usage{
JSTORplotVocabCount(df.list, legend, smooth = FALSE, YearLessThan)
}
\arguments{
\item{df.list}{A list or a single dataframe in the JSTOR_df format (see: JSTOR_df())}

\item{legend}{A vector of charcters to be used as legends matching
the number of df in df.list c("legend1", "legend2", ..., "legendN")}

\item{smooth}{Whether to smooth curves. Default smooth = FALSE}

\item{YearLessThan}{Upper bound for the years being ploted}
}
\value{
A dataframe of counted words JSTORVocabCount
}
\description{
This function plots the number of appearences of a set of words in abstracts or titles of a tidyJSTOR dataframe
}
\examples{
JSTORplotVocabCount(df, c("theory", "theories", "theoretical"))

JSTORplotVocabCount(c(df1, df2), c("df1Legend", "df2Legend"), smooth =TRUE, 2010)

}
