% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{JSTORrepeatedTopwords}
\alias{JSTORrepeatedTopwords}
\title{JSTORrepeatedTopwords}
\usage{
JSTORrepeatedTopwords(
  dfEco,
  y,
  x,
  StopWords = TRUE,
  output = "plot",
  YearBounds
)
}
\arguments{
\item{dfEco}{A dataframe in the JSTOR_df format (see: JSTOR_df())}

\item{y}{Determines the size of the ranking being observed for each year}

\item{x}{Determines the number of times a words has to be part of the rank to
be included in the dataframe or plot}

\item{StopWords}{Whether to use default StopWords or a custom dataframe of stop words. If opted
for custom, a dataframe of two columns (word, lexicon = "custom") must be provided.}

\item{output}{Whether to output a dataframe or a plot. Acceptable entries: "plot" and "dataframe".}

\item{YearBounds}{Optional vector with inferior and superior Year bounds for the plot. Ex: c(1970,1990)}
}
\value{
A ggplot2 object or a dataframe
}
\description{
Creates dataframe or plot with information regarding top words that appear at least
for a determined time period in the top.
}
\examples{
JSTORrepeatedTopwords(df, 3, 3)


}
