% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_search}
\alias{plot_search}
\title{plot_search}
\usage{
plot_search(df.list, legend, smooth = FALSE, YearLessThan)
}
\arguments{
\item{df.list}{A list or a single dataframe in the JSTOR_df format (see: JSTOR_df())}

\item{legend}{A vector of charcters to be used as legends matching
the number of df in df.list c("legend1", "legend2", ..., "legendN")}

\item{smooth}{Whether to smooth curves. Default smooth = FALSE}

\item{YearLessThan}{Upper bound for the years being ploted}
}
\value{
A plot of counted documents per year for a JSTOR_df
}
\description{
Plots number of documents per year in JSTOR_df after using SearchCount()
}
\examples{

}
