% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plotCleaningResult}
\alias{plotCleaningResult}
\title{plotCleaningResult}
\usage{
plotCleaningResult(dfClean, dfDirt)
}
\arguments{
\item{dfClean}{A dataframe in the JSTOR_df format (see: JSTOR_df()) after using CleanJSTOR_df()}

\item{dfDirt}{A dataframe in the JSTOR_df format (see: JSTOR_df())}
}
\value{
A plot
}
\description{
Plots number of documents per year in two distinct dataframes. Intended to use iwth a cleaned JSTOR_df and an original JSTOR_df
}
\examples{
plotCleaningResult(df1, df2)


}
