% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{SearchCount}
\alias{SearchCount}
\title{SearchCount}
\usage{
SearchCount(df)
}
\arguments{
\item{df}{A dataframe in the JSTOR_df format (see: JSTOR_df())}
}
\value{
A dataframe of counted documents per year for a JSTOR_df
}
\description{
Creates a dataframe with counted number of documents for each year in a JSTOR_df
}
\examples{

}
