% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyJSTOR.R
\name{Transform}
\alias{Transform}
\title{Transform}
\usage{
Transform(filePath)
}
\arguments{
\item{filePath}{}
}
\value{
A single row dataframe containing informations about: Journal, Publisher, Title, Year, Abstract and Language
}
\description{
This function transforms JSTOR DFR files into single row tidy dataframes
}
\examples{
Transform("C:/Users/.../Documents/metadata/file1.xml")


}
