% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{VocabCount}
\alias{VocabCount}
\title{JSTORVocabCount}
\usage{
VocabCount(dfEco, ArrayVocab, titles = FALSE, StopWords = TRUE, scores)
}
\arguments{
\item{dfEco}{A dataframe in the JSTOR_df format (see: JSTOR_df())}

\item{ArrayVocab}{a vector of charcters ot be counted as c("word1", "word2", ..., "wordn")}

\item{titles}{Whether to analyze titles of abstracts. Default titles = FALSE}

\item{StopWords}{Whether to use default StopWords or a custom dataframe of stop words. If opted
for custom, a dataframe of two columns (word, lexicon = "custom") must be provided.}

\item{scores}{a vector of scores matching the vector of character in the parameter ArrayVocab}
}
\value{
A dataframe of counted words JSTORVocabCount
}
\description{
This function counts the number of appearences of a set of words in abstracts or titles of a tidyJSTOR dataframe
}
\examples{
JSTORplotVocabCount(df, c("theory", "theories", "theoretical"))

JSTORVocabCount(df, c("theory", "theories", "theoretical"), scores = c(1,2,3))

}
